// frontend/web/src/api/types.gen/RouteDefinition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RouteBranch } from "./RouteBranch";
import type { RouteContextField } from "./RouteContextField";
import type { RouteGraphEdge } from "./RouteGraphEdge";
import type { RouteTraceMode } from "./RouteTraceMode";

export type RouteDefinition = { router_role: string, branches: Array<RouteBranch>, graph_edges?: Array<RouteGraphEdge>, context_fields: Array<RouteContextField>, trace_mode: RouteTraceMode, };